+++ example_test_suites/example_opensuse/case1/sakuli_demo.js
@@ -21,7 +21,7 @@ var testCase = new TestCase(60, 70);
 var env = new Environment();
 var screen = new Region();
 var appCalc = new Application("/usr/bin/kcalc");
-var appGedit = new Application("/usr/bin/kwrite");
+var appKwrite = new Application("/usr/bin/kwrite");
 
 try {
     _highlight(_link("SSL Manager"));
@@ -29,24 +29,31 @@ try {
     _highlight(_link("Online Documentation"));
     _highlight(_link("Test Pages"));
     _highlight(_link("Sample Application"));
-    
+
+    testCase.endOfStep("Test Sahi landing page",30);
     appCalc.open();
     screen.waitForImage("calculator.png", 20);
 
     env.type("525");
-    env.sleep(2);
-    screen.find("plus.png").click().type("100");
+    env.sleep(1);
+    screen.find("plus.png").click();
+    env.sleep(1);
+    env.type("100");
+    env.sleep(1);
     screen.find("result.png").click();
+    screen.waitForImage("625",10);
+    testCase.endOfStep("Calculation",30);
 
-    appGedit.open();
+    appKwrite.open();
     screen.waitForImage("kwrite.png", 20);
     env.paste("Initial test passed. Sakuli, Sahi and Sikuli seem to work fine. Exiting...");
+    testCase.endOfStep("Editor",30);
     env.sleep(4);
 } catch (e) {
     testCase.handleException(e);
 } finally {
-    appCalc.closeApp();
-    appGedit.closeApp();
+    appCalc.close();
+    appKwrite.close();
     if (screen.exists("close-without-saving", 1)) {
         screen.find("close-without-saving").click();
     }