+++ example_test_suites/example_windows/demo_win7/sakuli_demo.js
@@ -30,27 +30,31 @@ try {
     _highlight(_link("Test Pages"));
     _highlight(_link("Sample Application"));
 
+    testCase.endOfStep("Test Sahi landing page",30);
     appCalc.open();
     screen.waitForImage("calculator.png", 20);
 
     env.type("525");
-    env.sleep(2);
-    screen.find("plus.png").click().type("100");
-    screen.find("result.png").click();
+    env.sleep(1);
+    screen.find("plus.png").highlight().click().type("100");
+    screen.find("result.png").highlight().click();
+	screen.waitForImage("625",10).highlight();
+    testCase.endOfStep("Calculation",30);
 
     appNotepad.open();
-    screen.waitForImage("notepad.png", 20);
+    screen.waitForImage("notepad.png", 20).highlight();
     env.paste("Initial test passed. Sakuli, Sahi and Sikuli seem to work fine. Exiting...");
-    env.sleep(4);
+    testCase.endOfStep("Editor",30);
+    env.sleep(2);
 } catch (e) {
     testCase.handleException(e);
 } finally {
     testCase.saveResult();
     try {
-        appCalc.closeApp();
-        appNotepad.closeApp();
+        appCalc.close();
+        appNotepad.close();
     } catch (e) {
         //do not handle exception, this is just a teardown
-        env.logInfo(e);
+        Logger.logInfo(e);
     }
 }