+++ src/common/src/main/resources/org/sakuli/common/setup/database/create_sakuli_database.sql
@@ -8,9 +8,9 @@ CREATE SCHEMA IF NOT EXISTS `sakuli`
 USE `sakuli`;
 
 -- -----------------------------------------------------
--- Table `sakuli`.`sahi_suites`
+-- Table `sakuli`.`sakuli_suites`
 -- -----------------------------------------------------
-CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_suites` (
+CREATE TABLE IF NOT EXISTS `sakuli`.`sakuli_suites` (
   `id`          INT(11)       NOT NULL AUTO_INCREMENT,
   `suiteID`     VARCHAR(255)  NOT NULL,
   `result`      INT(11)       NOT NULL,
@@ -31,15 +31,15 @@ CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_suites` (
   UNIQUE INDEX `guid_UNIQUE` (`guid` ASC))
   ENGINE = MyISAM
   DEFAULT CHARACTER SET = utf8
-  COMMENT = 'Sahi Testcases';
+  COMMENT = 'Sakuli Testcases';
 
 
 -- -----------------------------------------------------
--- Table `sakuli`.`sahi_cases`
+-- Table `sakuli`.`sakuli_cases`
 -- -----------------------------------------------------
-CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_cases` (
+CREATE TABLE IF NOT EXISTS `sakuli`.`sakuli_cases` (
   `id`             INT(11)       NOT NULL AUTO_INCREMENT,
-  `sahi_suites_id` INT(11)       NOT NULL,
+  `sakuli_suites_id` INT(11) NOT NULL,
   `caseID`         VARCHAR(255)  NOT NULL,
   `result`         INT(11)       NOT NULL,
   `result_desc`    VARCHAR(45)   NOT NULL,
@@ -54,17 +54,18 @@ CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_cases` (
   `screenshot`     MEDIUMBLOB    NULL DEFAULT NULL,
   `msg`            VARCHAR(2500) NULL DEFAULT NULL,
   `time`           TIMESTAMP     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`, `sahi_suites_id`),
-  INDEX `fk_sahi_cases_sahi_suites` (`sahi_suites_id` ASC))
+  PRIMARY KEY (`id`, `sakuli_suites_id`),
+  INDEX `fk_sakuli_cases_sakuli_suites` (`sakuli_suites_id` ASC)
+)
   ENGINE = MyISAM
   DEFAULT CHARACTER SET = utf8
-  COMMENT = 'Sahi Testcases';
+  COMMENT = 'Sakuli Testcases';
 
 
 -- -----------------------------------------------------
--- Table `sakuli`.`sahi_jobs`
+-- Table `sakuli`.`sakuli_jobs`
 -- -----------------------------------------------------
-CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_jobs` (
+CREATE TABLE IF NOT EXISTS `sakuli`.`sakuli_jobs` (
   `id`   INT(11)      NOT NULL AUTO_INCREMENT,
   `guid` VARCHAR(255) NOT NULL,
   PRIMARY KEY (`id`),
@@ -74,11 +75,11 @@ CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_jobs` (
 
 
 -- -----------------------------------------------------
--- Table `sakuli`.`sahi_steps`
+-- Table `sakuli`.`sakuli_steps`
 -- -----------------------------------------------------
-CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_steps` (
+CREATE TABLE IF NOT EXISTS `sakuli`.`sakuli_steps` (
   `id`            INT(11)      NOT NULL AUTO_INCREMENT,
-  `sahi_cases_id` INT(11)      NOT NULL,
+  `sakuli_cases_id` INT(11) NOT NULL,
   `result`        INT(11)      NOT NULL,
   `result_desc`   VARCHAR(45)  NOT NULL,
   `name`          VARCHAR(255) NOT NULL,
@@ -87,11 +88,12 @@ CREATE TABLE IF NOT EXISTS `sakuli`.`sahi_steps` (
   `warning`       INT(11)      NULL DEFAULT NULL,
   `duration`      FLOAT        NOT NULL,
   `time`          TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`, `sahi_cases_id`),
-  INDEX `fk_sahi_steps_sahi_cases1` (`sahi_cases_id` ASC))
+  PRIMARY KEY (`id`, `sakuli_cases_id`),
+  INDEX `fk_sakuli_steps_sakuli_cases1` (`sakuli_cases_id` ASC)
+)
   ENGINE = MyISAM
   DEFAULT CHARACTER SET = utf8
-  COMMENT = 'Sahi Testcases';
+  COMMENT = 'Sakuli Testcases';
 
 
 SET SQL_MODE = @OLD_SQL_MODE;